feat: move some logic to be in custom hooks and remove fit on canva for pre-load on root graph
